processing_service/requirements.txt
Flask
python-dotenv==1.0.0
PyJWT==2.3.0
cryptography
requests